feat: add configuration and logo to printing
