+++ scripts/entrypoint.sh
@@ -9,13 +9,32 @@ mkdir -p /app/frontend/public
 node - <<'NODE'
 const fs = require('fs');
 const outPath = '/app/frontend/public/runtime-config.json';
-const keys = ['WIFI_SSID','WIFI_PASSWORD','WIFI_TYPE','WIFI_HIDDEN'];
+const keys = ['WIFI_SSID','WIFI_PASSWORD','WIFI_TYPE','WIFI_HIDDEN','PRINT_CONFIG'];
+
+function parseValue(value) {
+  if (!value) {
+    return value;
+  }
+
+  const trimmed = value.trim();
+
+  // Only attempt JSON parsing for object/array-looking values
+  if (trimmed.startsWith('{') || trimmed.startsWith('[')) {
+    try {
+      return JSON.parse(trimmed);
+    } catch {
+      // Not valid JSON, keep original string
+    }
+  }
+
+  return value;
+}
 
 const cfg = {};
 for (const k of keys) {
   const v = process.env[k];
   if (v !== undefined) {
-    cfg[k] = v;
+    cfg[k] = parseValue(v);
   }
 }
